requirements.txt
discord.py[voice]
peewee
tzdata